# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StudioFlow - Real Estate Media Delivery Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/media:
    get:
      operationId: listMedia
      tags: [media]
      summary: List media assets for a project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
        - in: query
          name: type
          schema:
            type: string
          required: false
        - in: query
          name: status
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of media assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaAsset"
    post:
      operationId: uploadMedia
      tags: [media]
      summary: Upload media asset (mock)
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadMediaBody"
      responses:
        "201":
          description: Media uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaAsset"
  /media/{id}:
    get:
      operationId: getMedia
      tags: [media]
      summary: Get media asset
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Media asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaAsset"
    patch:
      operationId: updateMedia
      tags: [media]
      summary: Update media asset
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMediaBody"
      responses:
        "200":
          description: Updated media
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaAsset"
    delete:
      operationId: deleteMedia
      tags: [media]
      summary: Delete media asset
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /media/{id}/approve:
    patch:
      operationId: approveMedia
      tags: [media]
      summary: Toggle media approval status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated media
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaAsset"

  # AI Jobs
...
  /projects/{projectId}/galleries:
    get:
      operationId: listGalleries
      tags: [galleries]
      summary: List galleries for a project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Galleries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Gallery"
    post:
      operationId: createGallery
      tags: [galleries]
      summary: Create gallery / share link
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGalleryBody"
      responses:
        "201":
          description: Gallery created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gallery"
  /galleries/{id}:
    get:
      operationId: getGallery
      tags: [galleries]
      summary: Get gallery
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Gallery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gallery"
    patch:
      operationId: updateGallery
      tags: [galleries]
      summary: Update gallery settings
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGalleryBody"
      responses:
        "200":
          description: Updated gallery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gallery"
    delete:
      operationId: deleteGallery
      tags: [galleries]
      summary: Delete gallery
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /gallery/{token}:
    get:
      operationId: getPublicGallery
      tags: [galleries]
      summary: Get public gallery by share token
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public gallery with assets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicGallery"
        "404":
          description: Not found

  # Comments
...
  /media/{mediaId}/favorite:
    post:
      operationId: toggleFavorite
      tags: [favorites]
      summary: Toggle favorite on a media asset
      parameters:
        - in: path
          name: mediaId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleFavoriteBody"
      responses:
        "200":
          description: Favorite toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteStatus"
  /projects/{projectId}/favorites:
    get:
      operationId: listFavorites
      tags: [favorites]
      summary: List favorited assets for a project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Favorited media assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaAsset"

  # Clients
...
components:
  schemas:
...
    MediaAsset:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        filename:
          type: string
        originalUrl:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        processedUrl:
          type: ["string", "null"]
        mediaType:
          type: string
          enum: [photo, video, virtual_tour]
        mimeType:
          type: string
        sizeMb:
          type: number
        width:
          type: ["integer", "null"]
        height:
          type: ["integer", "null"]
        durationSeconds:
          type: ["number", "null"]
        status:
          type: string
          enum: [uploading, ready, processing, processed, approved, rejected]
        isApproved:
          type: boolean
        isFavorited:
          type: boolean
        sortOrder:
          type: integer
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - projectId
        - filename
        - originalUrl
        - mediaType
        - mimeType
        - sizeMb
        - status
        - isApproved
        - isFavorited
        - sortOrder
        - tags
        - createdAt
        - updatedAt
    UploadMediaBody:
      type: object
      properties:
        filename:
          type: string
        originalUrl:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        mediaType:
          type: string
          enum: [photo, video, virtual_tour]
        mimeType:
          type: string
        sizeMb:
          type: number
        width:
          type: ["integer", "null"]
        height:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
      required:
        - filename
        - originalUrl
        - mediaType
        - mimeType
        - sizeMb
    UpdateMediaBody:
      type: object
      properties:
        filename:
          type: string
        tags:
          type: array
          items:
            type: string
        sortOrder:
          type: integer
        processedUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [uploading, ready, processing, processed, approved, rejected]
...
    Gallery:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        shareToken:
          type: string
        isPublic:
          type: boolean
        isPasswordProtected:
          type: boolean
        allowDownload:
          type: boolean
        allowFavorites:
          type: boolean
        allowComments:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        viewCount:
          type: integer
        downloadCount:
          type: integer
        mediaIds:
          type: array
          items:
            type: integer
        clientMessage:
          type: ["string", "null"]
        brandingLogoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - projectId
        - name
        - shareToken
        - isPublic
        - isPasswordProtected
        - allowDownload
        - allowFavorites
        - allowComments
        - viewCount
        - downloadCount
        - mediaIds
        - createdAt
    CreateGalleryBody:
      type: object
      properties:
        name:
          type: string
        isPublic:
          type: boolean
        isPasswordProtected:
          type: boolean
        password:
          type: ["string", "null"]
        allowDownload:
          type: boolean
        allowFavorites:
          type: boolean
        allowComments:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        mediaIds:
          type: array
          items:
            type: integer
        clientMessage:
          type: ["string", "null"]
      required:
        - name
    UpdateGalleryBody:
      type: object
      properties:
        name:
          type: string
        isPublic:
          type: boolean
        isPasswordProtected:
          type: boolean
        password:
          type: ["string", "null"]
        allowDownload:
          type: boolean
        allowFavorites:
          type: boolean
        allowComments:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        mediaIds:
          type: array
          items:
            type: integer
        clientMessage:
          type: ["string", "null"]
    PublicGallery:
      type: object
      properties:
        id:
          type: integer
        projectName:
          type: string
        projectAddress:
          type: string
        galleryName:
          type: string
        allowDownload:
          type: boolean
        allowFavorites:
          type: boolean
        allowComments:
          type: boolean
        clientMessage:
          type: ["string", "null"]
        brandingLogoUrl:
          type: ["string", "null"]
        photographerName:
          type: string
        media:
          type: array
          items:
            $ref: "#/components/schemas/MediaAsset"
      required:
        - id
        - projectName
        - projectAddress
        - galleryName
        - allowDownload
        - allowFavorites
        - allowComments
        - photographerName
        - media
...
    FavoriteStatus:
      type: object
      properties:
        mediaId:
          type: integer
        isFavorited:
          type: boolean
      required:
        - mediaId
        - isFavorited
    ToggleFavoriteBody:
      type: object
      properties:
        clientName:
          type: ["string", "null"]
        galleryToken:
          type: ["string", "null"]
...