# lib/api-zod/src/generated/types/favoriteStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudioFlow - Real Estate Media Delivery Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface FavoriteStatus {
  mediaId: number;
  isFavorited: boolean;
}